requirements.txt
PyQt6
pillow
requests